Added bookmarks and last read feature
Changed book class: Added bookmarklist and lastread.
Corrected nullable data types wherever needed.
Dynamically adds bookmarks to the list which is visible on the appbar
of pdf viewer by clicking on the bookmark button.

How it works?
Click on the bookmark icon to add the bookmark of the page user is on.
Move either to the next page, previous page, refersh or the main screen
to make it visible in the above bookmark bar.
It gets updated but is not visible without using any of the above buttons.

What can be fixed?

1. Dual entries of bookmarks are allowed. Need to fix that. Can be fixed
while working on dual tabs.
2. The bookmark box which has the page number appears only when some other button is clicked (forward, backward, go to main screen or refresh).
3. Dispose wasn't working out so didn't use that.
4. Didn't test this but pls check if the lastpage read is retained on closing the application.
5. Pls check the size of the bar above and also the viewer size.