Added option to choose initial page
On clicking the book, pop up opens and the user gets an option to select where to start reading: From beginning or From the point he left off.
There might be some null check errors on console because of the parameters used. If that can be fixed, great! Otherwise it isnt causing any issue in the application anyway.